# Optimize EXIT, C!, C@ and 0=